Connect RootPage
