+++ app/src/components/GenericSubApp/GenericSubAppProps.ts
@@ -1,8 +1,8 @@
+import { SideMenuProps } from 'components/GenericSubApp/SideMenuProps';
 import { FC } from 'react';
 import { Store } from 'redux';
 import { SnowmanAction } from 'store/messages';
 import { ViewIDs } from 'types/ViewIDs';
-import { SideMenuProps } from 'types/ViewMetaInformation';
 
 export interface GenericSubAppStateProps {
   existsActiveRequest: boolean;
