+++ app/src/components/GenericSubApp/SideMenuProps.ts
@@ -1,0 +1,3 @@
+export interface SideMenuProps {
+  contentId: string;
+}
